fix: PR review fixes — weak test, debug logging, trace.py rename [skip review]
- Remove always-passing fallback in test_compose_core_only
- Add logger.debug to _get_node_interface exception handler
- Rename trace.py → report.py to match the command it exports
- Add pflow report mention to entry.md error guidance
