+++ src/pflow/cli/commands/CLAUDE.md
@@ -18,7 +18,7 @@ One file per top-level command, registered in `main.py` via `cli.add_command()`.
 | `read_fields.py` | `pflow read-fields <exec> <paths>` | `pflow.core.execution_cache` |
 | `skills.py` | `pflow skill save\|list\|remove` | `pflow.core.workflow.skill_service` |
 | `settings.py` | `pflow settings ...` | `pflow.core.settings` |
-| `trace.py` | `pflow report` | `pflow.core.trace_report` |
+| `report.py` | `pflow report` | `pflow.core.trace_report` |
 | `visualize.py` | `pflow visualize` | `pflow.core.workflow.mermaid`, `pflow.execution.*` |
 
 ## Cross-References Within commands/
